requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
astroid==2.7.3
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==21.0
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.7.0
pylint==2.10.2
pyparsing==2.4.7
pytest==6.2.5
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml.clib==0.2.6
six==1.16.0
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
